move psi u64 types to i32

+++ crates/scouter_drift/src/psi/drift.rs
@@ -234,7 +234,7 @@ pub mod psi_drifter {
         fn create_feature_bin_proportion_pairs(
             &self,
             feature: &str,
-            bin_proportions: &BTreeMap<usize, f64>,
+            bin_proportions: &BTreeMap<i32, f64>,
         ) -> Result<FeatureBinProportionPairs, DriftError> {
             // get profile
             let profile = match self.profile.features.get(feature) {
